image-analysis/pyproject.toml
[project]
name = "test-any"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "exif>=1.6.1",
    "exifread==3.5.1",
    "opencv-python>=4.12.0.88",
    "pillow>=10.4.0",
    "pillow-heif>=0.18.0",
]
